move state/tstate, keys and pubsub out of internal
