+++ pubsub/errors.go
