+++ state/tstate/tstate_test.go
@@ -15,8 +15,8 @@ import (
 	"github.com/ava-labs/avalanchego/x/merkledb"
 	"github.com/stretchr/testify/require"
 
-	"github.com/ava-labs/hypersdk/internal/keys"
 	"github.com/ava-labs/hypersdk/internal/trace"
+	"github.com/ava-labs/hypersdk/keys"
 	"github.com/ava-labs/hypersdk/state"
 )
